requirements.txt
fastapi
requests
python-dateutil
uvicorn[standard]
